requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langgraph>=0.0.50
openai>=1.30.0
chromadb>=0.5.0
tavily-python>=0.3.4
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
pytest>=8.1.0
